pyproject.toml
[tool.poetry]
name = "maestro"
version = "1.0.0"
description = "Manage customers, tasks and invoices."
authors = ["Kevin Bieri <bierike@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
django-heroku = "^0.3.1"
djangorestframework = "^3.12.4"
django-storages = "^1.11.1"
boto3 = "^1.17.49"
gunicorn = "^20.1.0"
django-extensions = "^3.1.2"
pluck = "^0.2"
django-colorfield = "^0.4.1"
django-recurrence = "^1.10.3"
Jinja2 = "^2.11.3"
arrow = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
flake8 = "^3.9.0"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
